requirements-manual.txt
discord.py
GitPython
pyparsing
hashids
pytz
